Adding methods for short / ushort

+++ src/RndF/Functions/Rnd.NumberF.GetInt16.cs
@@ -11,14 +11,14 @@ public static partial class Rnd
 	public static partial class NumberF
 	{
 		/// <summary>
-		/// Returns a random positive integer between <see langword="0"/> and <see cref="int.MaxValue"/> inclusive.
+		/// Returns a random positive integer between <see langword="0"/> and <see cref="short.MaxValue"/> inclusive.
 		/// </summary>
 		/// <remarks>
 		/// Don't share code with other GetIntxx() methods for memory allocation reasons.
 		/// </remarks>
 		/// <returns>Random number.</returns>
-		public static int GetInt32() =>
-			GetInt32(0, int.MaxValue);
+		public static short GetInt16() =>
+			GetInt16(0, short.MaxValue);
 
 		/// <summary>
 		/// Returns a random positive integer between <see langword="0"/> and <paramref name="max"/> inclusive.
@@ -29,8 +29,8 @@ public static partial class Rnd
 		/// <param name="max">Maximum acceptable value.</param>
 		/// <returns>Random number.</returns>
 		/// <exception cref="MaximumLessThanMinimumException"/>
-		public static int GetInt32(int max) =>
-			GetInt32(0, max);
+		public static short GetInt16(short max) =>
+			GetInt16(0, max);
 
 		/// <summary>
 		/// Returns a random positive integer between <paramref name="min"/> and <paramref name="max"/> inclusive.
@@ -43,17 +43,17 @@ public static partial class Rnd
 		/// <returns>Random number.</returns>
 		/// <exception cref="MaximumLessThanMinimumException"/>
 		/// <exception cref="MinimumLessThanZeroException"/>
-		public static int GetInt32(int min, int max)
+		public static short GetInt16(short min, short max)
 		{
 			// Check arguments
 			if (min >= max)
 			{
-				throw MaximumLessThanMinimumException.Create(nameof(GetInt32), min, max);
+				throw MaximumLessThanMinimumException.Create(nameof(GetInt16), min, max);
 			}
 
 			if (min < 0)
 			{
-				throw MinimumLessThanZeroException.Create(nameof(GetInt32), min);
+				throw MinimumLessThanZeroException.Create(nameof(GetInt16), min);
 			}
 
 			// Get the range between the specified minimum and maximum values
@@ -61,7 +61,7 @@ public static partial class Rnd
 
 			// Now add a random amount of the range to the minimum value - it will never exceed maximum value
 			var add = Math.Round(range * Get());
-			return (int)(min + add);
+			return (short)(min + add);
 		}
 	}
 }
